refactor: folder filter list intro separate components
