+++ packages/react-ui/src/features/folders/component/folder-filter-list/index.tsx
@@ -1,12 +1,7 @@
 import { PlusIcon } from '@radix-ui/react-icons';
 import { useQuery } from '@tanstack/react-query';
 import { t } from 'i18next';
-import {
-  ChevronDown,
-  EllipsisVertical,
-  Pencil,
-  Trash2,
-} from 'lucide-react';
+import { ChevronDown } from 'lucide-react';
 import { useEffect, useMemo, useState } from 'react';
 import { useLocation, useSearchParams } from 'react-router-dom';
 import {
@@ -17,11 +12,9 @@ import {
   useSensor,
   useSensors,
 } from '@dnd-kit/core';
-import { SortableContext, horizontalListSortingStrategy, useSortable } from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
+import { SortableContext, horizontalListSortingStrategy } from '@dnd-kit/sortable';
 import { restrictToHorizontalAxis } from '@dnd-kit/modifiers';
 
-import { ConfirmationDeleteDialog } from '@/components/delete-dialog';
 import { Button } from '@/components/ui/button';
 import {
   Command,
@@ -30,11 +23,6 @@ import {
   CommandGroup,
   CommandItem,
 } from '@/components/ui/command';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
 import { PermissionNeededTooltip } from '@/components/ui/permission-needed-tooltip';
 import {
   Popover,
@@ -48,196 +36,15 @@ import { authenticationSession } from '@/lib/authentication-session';
 import { cn } from '@/lib/utils';
 import { FolderDto, isNil, Permission } from '@activepieces/shared';
 
-import { foldersApi } from '../lib/folders-api';
-import { foldersHooks } from '../lib/folders-hooks';
-import { foldersUtils } from '../lib/folders-utils';
+import { foldersHooks } from '../../lib/folders-hooks';
+import { foldersUtils } from '../../lib/folders-utils';
 
-import { CreateFolderDialog } from './create-folder-dialog';
-import { RenameFolderDialog } from './rename-folder-dialog';
+import { CreateFolderDialog } from '../create-folder-dialog';
+import { FolderAction } from './folder-action';
+import { SortableFolder } from './sortable-folder';
 
-const folderIdParamName = 'folderId';
+export const folderIdParamName = 'folderId';
 const FOLDER_ORDER_STORAGE_KEY = 'ap_folder_order';
-
-type FolderActionProps = {
-  folder: FolderDto;
-  refetch: () => void;
-  userHasPermissionToUpdateFolders: boolean;
-};
-
-export const FolderAction = ({
-  folder,
-  refetch,
-  userHasPermissionToUpdateFolders,
-}: FolderActionProps) => {
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [dialogOpen, setDialogOpen] = useState<'rename' | 'delete' | null>(
-    null,
-  );
-
-  const handleOpenDialog = (dialog: 'rename' | 'delete') => {
-    setDialogOpen(dialog);
-    setIsDropdownOpen(false);
-  };
-
-  const handleDialogClose = () => {
-    setDialogOpen(null);
-  };
-
-  return (
-    <>
-      <DropdownMenu open={isDropdownOpen} onOpenChange={setIsDropdownOpen}>
-        <DropdownMenuTrigger asChild>
-          <Button variant="ghost" className="h-9" size="icon">
-            <EllipsisVertical className="h-4 w-4" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent>
-          <PermissionNeededTooltip
-            hasPermission={userHasPermissionToUpdateFolders}
-          >
-            <Button
-              variant="ghost"
-              size="sm"
-              className="flex h-8 w-full justify-start gap-2 items-center"
-              onClick={(e) => {
-                e.stopPropagation();
-                handleOpenDialog('rename');
-              }}
-            >
-              <Pencil className="h-4 w-4" />
-              <span>{t('Rename')}</span>
-            </Button>
-          </PermissionNeededTooltip>
-          <PermissionNeededTooltip
-            hasPermission={userHasPermissionToUpdateFolders}
-          >
-            <Button
-              variant="ghost"
-              size="sm"
-              className="flex h-8 w-full justify-start gap-2 items-center"
-              onClick={(e) => {
-                e.stopPropagation();
-                handleOpenDialog('delete');
-              }}
-            >
-              <Trash2 className="h-4 w-4 text-destructive" />
-              <span className="text-destructive">{t('Delete')}</span>
-            </Button>
-          </PermissionNeededTooltip>
-        </DropdownMenuContent>
-      </DropdownMenu>
-
-      {dialogOpen === 'rename' && (
-        <RenameFolderDialog
-          folderId={folder.id}
-          name={folder.displayName}
-          onRename={refetch}
-          open={true}
-          onOpenChange={(open) => {
-            if (!open) handleDialogClose();
-          }}
-        />
-      )}
-
-      {dialogOpen === 'delete' && (
-        <ConfirmationDeleteDialog
-          title={t('Delete {folderName}', {
-            folderName: folder.displayName,
-          })}
-          message={t(
-            'If you delete this folder, we will keep its flows and move them to Uncategorized.',
-          )}
-          mutationFn={async () => {
-            console.info('HEllo');
-            await foldersApi.delete(folder.id);
-            refetch();
-          }}
-          entityName={folder.displayName}
-          open={true}
-          onOpenChange={(open) => {
-            if (!open) handleDialogClose();
-          }}
-        />
-      )}
-    </>
-  );
-};
-
-// Sortable folder component
-const SortableFolder = ({
-  folder,
-  isSelected,
-  onClick,
-  refetch,
-  userHasPermissionToUpdateFolders,
-}: {
-  folder: FolderDto;
-  isSelected: boolean;
-  onClick: () => void;
-  refetch: () => void;
-  userHasPermissionToUpdateFolders: boolean;
-}) => {
-  const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
-    id: folder.id,
-  });
-  
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-    zIndex: isDragging ? 100 : 1,
-  };
-
-  const [emoji, ...nameParts] = folder.displayName.split(' ');
-  const name = nameParts.join(' ');
-
-  // Handle container click to avoid interfering with button clicks
-  const handleContainerClick = (e: React.MouseEvent) => {
-    // Only apply drag listeners to container directly, not to child elements
-    if (e.target === e.currentTarget) {
-      // This is a direct click on the container, not on a child
-      e.preventDefault();
-    }
-  };
-
-  return (
-    <div 
-      ref={setNodeRef} 
-      style={style}
-      {...attributes}
-      {...listeners}
-      onClick={handleContainerClick}
-      className={cn(
-        "relative group whitespace-nowrap h-9 flex overflow-hidden items-center border rounded-md cursor-grab active:cursor-grabbing",
-        isDragging ? "opacity-60 shadow-md" : "opacity-100",
-        isSelected ? "bg-secondary" : "bg-background"
-      )}
-    >
-      <Button
-        variant={isSelected ? 'secondary' : 'ghost'}
-        size="sm"
-        onClick={onClick}
-        className="group whitespace-nowrap flex overflow-hidden items-center px-3 border-0 z-10"
-      >
-        <span className="mr-2">{emoji}</span>
-        <span className="mr-2 flex items-center">
-          {name}
-          <span className="text-xs text-muted-foreground ml-1">
-            ({folder.numberOfFlows})
-          </span>
-        </span>
-      </Button>
-      <div>
-        <FolderAction
-          folder={folder}
-          refetch={refetch}
-          userHasPermissionToUpdateFolders={
-            userHasPermissionToUpdateFolders
-          }
-        />
-      </div>
-    </div>
-  );
-};
 
 const FolderFilterList = ({ refresh }: { refresh: number }) => {
   const location = useLocation();
@@ -245,10 +52,7 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
   const userHasPermissionToUpdateFolders = checkAccess(Permission.WRITE_FOLDER);
   const [searchParams, setSearchParams] = useSearchParams(location.search);
   const selectedFolderId = searchParams.get(folderIdParamName);
-  const [
-    sortedAlphabeticallyIncreasingly,
-    setSortedAlphabeticallyIncreasingly,
-  ] = useState(true);
+  const [sortedAlphabeticallyIncreasingly, setSortedAlphabeticallyIncreasingly] = useState(true);
   const [showMoreFolders, setShowMoreFolders] = useState(false);
   const [isDragging, setIsDragging] = useState(false);
   
@@ -299,7 +103,6 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
     if (folderOrder.length > 0) {
       const orderMap = new Map(folderOrder.map((id, index) => [id, index]));
       
-      // Sort using the order map, folders not in the map go to the end
       return foldersCopy.sort((a, b) => {
         const aIndex = orderMap.has(a.id) ? orderMap.get(a.id)! : Number.MAX_SAFE_INTEGER;
         const bIndex = orderMap.has(b.id) ? orderMap.get(b.id)! : Number.MAX_SAFE_INTEGER;
@@ -307,7 +110,6 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
       });
     }
     
-    // Fall back to alphabetical sorting if no custom order
     return foldersCopy.sort((a, b) => {
       if (sortedAlphabeticallyIncreasingly) {
         return a.displayName.localeCompare(b.displayName);
@@ -317,7 +119,6 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
     });
   }, [folders, folderOrder, sortedAlphabeticallyIncreasingly]);
   
-  // Separate the ordered folders into visible and more
   const visibleFolders = useMemo(() => {
     return orderedFolders.slice(0, visibleFolderCount) || [];
   }, [orderedFolders, visibleFolderCount]);
@@ -331,7 +132,6 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
     refetchAllFlowsCount();
   }, [refresh]);
   
-  // Initialize folder order with existing folders if order is empty
   useEffect(() => {
     if (folders && folders.length > 0 && folderOrder.length === 0) {
       const newOrder = folders.sort((a, b) => 
@@ -346,7 +146,6 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
   const isInUncategorized = selectedFolderId === 'NULL';
   const isInAllFlows = isNil(selectedFolderId);
 
-  // Handle drag end
   const handleDragEnd = (event: DragEndEvent) => {
     setIsDragging(false);
     const { active, over } = event;
@@ -362,8 +161,10 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
       
       if (oldIndex !== -1 && newIndex !== -1) {
         const newOrder = [...currentIds];
+        
         newOrder.splice(oldIndex, 1);
         newOrder.splice(newIndex, 0, activeId);
+        
         setFolderOrder(newOrder);
         localStorage.setItem(FOLDER_ORDER_STORAGE_KEY, JSON.stringify(newOrder));
       }
@@ -376,6 +177,7 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
 
   const promoteFolder = (folderId: string) => {
     if (!folders) return;
+    
     const isAlreadyVisible = visibleFolders.some(folder => folder.id === folderId);
     if (isAlreadyVisible) return;
     
@@ -400,7 +202,7 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
           variant={isInAllFlows ? 'secondary' : 'ghost'}
           size="sm"
           onClick={() => updateSearchParams(undefined)}
-          className="group border"
+          className="group border h-9"
         >
           <span className="mr-2">🗂️</span>
           {t(`All`)}
@@ -413,7 +215,7 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
           variant={isInUncategorized ? 'secondary' : 'ghost'}
           size="sm"
           onClick={() => updateSearchParams('NULL')}
-          className="group border"
+          className="group border h-9"
         >
           <span className="mr-2">📦</span>
           {t('Uncategorized')}
@@ -444,7 +246,7 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
               <div className={cn(
                 "flex items-center gap-2 min-w-0",
                 isDragging 
-                  ? "overflow-x-auto pb-2 [&::-webkit-scrollbar]:hidden [-ms-overflow-style:none] [scrollbar-width:none] whitespace-nowrap" 
+                  ? "overflow-x-auto [&::-webkit-scrollbar]:hidden [-ms-overflow-style:none] [scrollbar-width:none] whitespace-nowrap" 
                   : "flex-wrap"
               )}>
                 {(isDragging ? orderedFolders : visibleFolders).map((folder) => (
@@ -553,4 +355,4 @@ const FolderFilterList = ({ refresh }: { refresh: number }) => {
   );
 };
 
-export { FolderFilterList, folderIdParamName };
+export { FolderFilterList };
